getting unclaimed reward for nomination pool
